Add register and login forms and functionality

+++ todo/routes.py
@@ -1,20 +1,25 @@
 from todo import app
-from flask import render_template, redirect, url_for
-from todo.models import Task
+from flask import render_template, redirect, url_for, flash
+from todo.models import Task, User
 from todo.forms import *
 from todo import db
+from flask_login import login_user, logout_user, current_user
 
-@app.route('/', methods=['GET'])
-@app.route('/home')
+@app.route('/')
+@app.route('/home', methods=['GET', 'POST'])
 def home_page():
-    add_task_form = AddTaskForm()
-    edit_task_form = EditTaskForm()
-    complete_task_form = CompleteTaskForm()
-    delete_task_form = DeleteTaskForm()
-    
-    tasks = Task.query.all()
-    return render_template('index.html', add_task_form=add_task_form, edit_task_form=edit_task_form,
-    complete_task_form=complete_task_form, delete_task_form=delete_task_form, tasks=tasks)
+    if current_user.is_authenticated:
+        add_task_form = AddTaskForm()
+        edit_task_form = EditTaskForm()
+        complete_task_form = CompleteTaskForm()
+        delete_task_form = DeleteTaskForm()
+        
+        tasks = Task.query.filter_by(user_id = current_user.get_id())
+
+        return render_template('index.html', add_task_form=add_task_form, edit_task_form=edit_task_form,
+        complete_task_form=complete_task_form, delete_task_form=delete_task_form, tasks=tasks)
+    else:
+        return redirect(url_for('login_page'))
 
 @app.route('/add', methods=['POST'])
 def add():
@@ -24,7 +29,7 @@ def add():
     delete_task_form = DeleteTaskForm()
 
     if add_task_form.validate_on_submit():
-        new_task = Task(description=add_task_form.description.data)
+        new_task = Task(user_id=current_user.get_id(), description=add_task_form.description.data)
         db.session.add(new_task)
         db.session.commit()
         return redirect(url_for('home_page'))
@@ -88,10 +93,45 @@ def delete():
     return render_template('index.html', add_task_form=add_task_form, edit_task_form=edit_task_form,
     complete_task_form=complete_task_form, delete_task_form=delete_task_form)
 
-@app.route('/register')
+@app.route('/register', methods=['GET', 'POST'])
 def register_page():
-    return render_template('register.html')
+    form = RegisterForm()
 
-@app.route('/login')
+    if form.validate_on_submit():
+        new_user = User(username=form.username.data,
+                        email_address=form.email_address.data,
+                        password=form.password.data)
+
+        db.session.add(new_user)
+        db.session.commit()
+        login_user(new_user)
+        flash(f'Welcome {new_user.username}!', category='success')
+        return redirect(url_for('home_page'))
+
+    if form.errors != {}:
+        for err_msg in form.errors.values():
+            flash(f'There was an error with creating the account: {err_msg}', category='danger')
+
+    return render_template('register.html', form=form)
+
+@app.route('/login', methods=['GET', 'POST'])
 def login_page():
-    return render_template('login.html')
+    form = LoginForm()
+
+    if form.validate_on_submit():
+        entered_user = User.query.filter_by(username=form.username.data).first()
+
+        if entered_user and entered_user.check_password(entered_password=form.password.data):
+            login_user(entered_user)
+            flash('Successfully logged in', category='success')
+            return redirect(url_for('home_page'))
+        else:
+            flash('Invalid username or password. Try again', category='danger')
+
+    return render_template('login.html', form=form)
+
+@app.route('/logout')
+def logout_page():
+    logout_user()
+    flash('You have been successfully logged out', category='info')
+    return redirect(url_for('login_page'))
